crosslayer: Move information into its own package
